requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
gunicorn==21.2.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
tensorflow-cpu==2.12.0
matplotlib==3.7.1
plotly==5.14.1
joblib==1.2.0